backend/api/requirements.txt
# backend/requirements.txt

fastapi
uvicorn
pandas
beautifulsoup4
psycopg2-binary
requests
selenium